(refactor) remove layersSize test. decoding and images should be sufficient plus...
Podman appears to possibly have a bug around the LayersSize key in the output json... Looking at the podman docs, there's no reference to LayerSize, but there is an ImagesSize, but that is also in the docs as `0`. It's definitely weird, so I'm just disabling layerSize since it appears to be giving a false flag failure.